Fix deletion provider and most tests to runtime
