+++ crates/data_components/Cargo.toml
@@ -78,9 +78,9 @@ rdkafka = { version = "0.36.2", features = ["cmake-build"], optional = true }
 clickhouse = ["dep:clickhouse-rs", "arrow_sql_gen/clickhouse"]
 databricks = ["dep:deltalake", "spark_connect"]
 debezium = ["dep:serde_json", "dep:rdkafka"]
-duckdb = ["dep:duckdb", "dep:r2d2", "db_connection_pool/duckdb"]
+duckdb = ["dep:duckdb", "dep:r2d2", "db_connection_pool/duckdb", "datafusion-table-providers/duckdb"]
 flightsql = ["dep:tonic", "dep:r2d2"]
-mysql = ["dep:mysql_async", "arrow_sql_gen/mysql"]
+mysql = ["dep:mysql_async", "arrow_sql_gen/mysql", "datafusion-table-providers/mysql"]
 odbc = ["dep:odbc-api", "dep:arrow-odbc"]
 postgres = [
   "dep:bb8",
@@ -92,7 +92,7 @@ postgres = [
 ]
 snowflake = ["dep:snowflake-api"]
 spark_connect = ["dep:spark-connect-rs"]
-sqlite = ["dep:rusqlite", "dep:tokio-rusqlite", "arrow_sql_gen/sqlite"]
+sqlite = ["dep:rusqlite", "dep:tokio-rusqlite", "arrow_sql_gen/sqlite", "datafusion-table-providers/sqlite"]
 
 [dev-dependencies]
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
